@@ -1,16 +1,16 @@
-# Be sure to restart your server when you modify this file.
+# frozen_string_literal: true
 
-# Avoid CORS issues when API is called from the frontend app.
-# Handle Cross-Origin Resource Sharing (CORS) in order to accept cross-origin AJAX requests.
+# Configuração de CORS para permitir requisições cross-origin
+# Necessário para aplicações frontend (React, Vue, etc.)
 
-# Read more: https://github.com/cyu/rack-cors
+Rails.application.config.middleware.insert_before 0, Rack::Cors do
+  allow do
+    # Em produção, altere para o domínio específico do frontend
+    origins '*'
 
-# Rails.application.config.middleware.insert_before 0, Rack::Cors do
-#   allow do
-#     origins 'example.com'
-#
-#     resource '*',
-#       headers: :any,
-#       methods: [:get, :post, :put, :patch, :delete, :options, :head]
-#   end
-# end
+    resource '*',
+             headers: :any,
+             expose: %w[access-token expiry token-type uid client], # Headers do Devise Token Auth
+             methods: %i[get post put patch delete options head]
+  end
+end
